+++ packages/create-app/templates/default-app/packages/app/src/App.test.tsx
@@ -4,9 +4,9 @@ import App from './App';
 
 describe('App', () => {
   it('should render', async () => {
-    Object.defineProperty(process.env, 'APP_CONFIG', {
-      configurable: true,
-      value: [
+    process.env = {
+      NODE_ENV: 'test',
+      APP_CONFIG: [
         {
           data: {
             app: { title: 'Test' },
@@ -17,8 +17,8 @@ describe('App', () => {
           },
           context: 'test',
         },
-      ],
-    });
+      ] as any,
+    };
 
     const rendered = await renderWithEffects(<App />);
     expect(rendered.baseElement).toBeInTheDocument();