+++ plugins/cost-insights/src/forms/index.ts
@@ -17,5 +17,3 @@
 export { AlertAcceptForm } from './AlertAcceptForm';
 export { AlertSnoozeForm } from './AlertSnoozeForm';
 export { AlertDismissForm } from './AlertDismissForm';
-export { MigrationDismissForm } from './MigrationDismissForm';
-export type { MigrationDismissFormData } from './MigrationDismissForm';