+++ plugins/splunk-on-call/src/components/SplunkOnCallCard.test.tsx
@@ -50,7 +50,7 @@ const mockSplunkOnCallApi: Partial<SplunkOnCallClient> = {
 
 const configApi: ConfigApi = new ConfigReader({
   splunkOnCall: {
-    username: MOCKED_USER.username,
+    eventsRestEndpoint: 'EXAMPLE_REST_ENDPOINT',
   },
 });
 
@@ -91,7 +91,10 @@ describe('SplunkOnCallCard', () => {
     );
     await waitFor(() => !queryByTestId('progress'));
     expect(getByText('Create Incident')).toBeInTheDocument();
-    expect(getByText('Nice! No incidents found!')).toBeInTheDocument();
+    await waitFor(
+      () => expect(getByText('Nice! No incidents found!')).toBeInTheDocument(),
+      { timeout: 2000 },
+    );
     expect(getByText('Empty escalation policy')).toBeInTheDocument();
   });