+++ plugins/github-actions/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @backstage/plugin-github-actions
 
+## 0.3.5
+
+### Patch Changes
+
+- b2a5320a4: Use Breadcrumbs from @backstage/core rather than material-ui
+- 04667f571: Use Breadcrumbs from @backstage/core
+- Updated dependencies [12d8f27a6]
+- Updated dependencies [40c0fdbaa]
+- Updated dependencies [2a271d89e]
+- Updated dependencies [bece09057]
+- Updated dependencies [169f48deb]
+- Updated dependencies [8a1566719]
+- Updated dependencies [9d455f69a]
+- Updated dependencies [4c049a1a1]
+- Updated dependencies [02816ecd7]
+  - @backstage/catalog-model@0.7.3
+  - @backstage/core@0.7.0
+  - @backstage/plugin-catalog-react@0.1.1
+
 ## 0.3.4
 
 ### Patch Changes